fix: resolve clippy warnings across benches and tests (29 files)
Fixes: mul_add, semicolon_if_nothing_returned, unnecessary_cast,
float_cmp, manual_find, redundant_closure, explicit_iter_loop,
wildcard_imports, if_same_then_else, manual_range_contains,
needless_range_loop, needless_collect, similar_names, io_other_error,
large_futures, unnecessary_get_then_check, unused imports/mut,
manual_checked_ops, suboptimal_flops, missing CompatRuntime imports.

+++ crates/frankenterm-core/benches/token_bucket_bench.rs
@@ -63,7 +63,7 @@ fn bursty_traffic(n: usize) -> Vec<(u32, u64)> {
         // Cost: 1-5 tokens
         let cost = 1 + ((state >> 60) as u32 % 5);
         // Inter-arrival: 1-200ms (bursty)
-        let interval = 1 + (state >> 50) as u64 % 200;
+        let interval = 1 + (state >> 50) % 200;
         t += interval;
         out.push((cost, t));
     }
